feat: fixing web build issues
