+++ package.json
@@ -109,6 +109,7 @@
     "@babel/preset-flow": "^7.16.7",
     "@digitalbazaar/x25519-key-agreement-key-2020": "2.1.0",
     "@react-native-async-storage/async-storage": "^1.15.17",
+    "@rollup/plugin-alias": "^4.0.2",
     "@rollup/plugin-babel": "^5.3.0",
     "@rollup/plugin-commonjs": "^21.0.1",
     "@rollup/plugin-json": "^4.1.0",
