+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@docknetwork/wallet-sdk-core",
-  "version": "0.3.7",
+  "version": "0.3.10",
   "license": "MIT",
   "repository": {
     "type": "git",
@@ -14,13 +14,14 @@
   "peerDependencies": {
     "typeorm": "^0.3.15",
     "reflect-metadata": "^0.1.13",
-    "sqlite3": "^5.0.2"
+    "sqlite3": "^5.0.2",
+    "@docknetwork/wallet-sdk-data-store": "^0.3.9"
   },
   "dependencies": {
-    "@docknetwork/wallet-sdk-wasm": "0.3.7",
-    "@docknetwork/wallet-sdk-data-store": "0.3.7"
+    "@docknetwork/wallet-sdk-wasm": "^0.3.8"
   },
   "devDependencies": {
+    "@docknetwork/wallet-sdk-data-store": "0.3.8",
     "jest": "29.1.0",
     "ts-jest": "29.1.0",
     "ts-node": "^10.9.1",
