+++ packages/data-store/src/migration/bootstrap-tables.ts
@@ -1,4 +1,4 @@
-import {DataSource} from 'typeorm';
+import {DataSource} from '../typeorm';
 
 export function bootstrapTables(dataSource: DataSource) {
   // will be generated from typeorm once the schema is defined
