+++ packages/wasm/babel.config.js
@@ -2,11 +2,7 @@ module.exports = {
   presets: [
     [
       '@babel/preset-env',
-      {
-        targets: {
-          node: 'current',
-        },
-      },
+
     ],
   ],
   plugins: [
