+++ packages/wasm/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@docknetwork/wallet-sdk-wasm",
-    "version": "0.3.7",
+    "version": "0.3.8",
     "license": "MIT",
     "repository": {
         "type": "git",
@@ -8,7 +8,8 @@
     },
     "scripts": {
         "prepublishOnly": "npm run build",
-        "build": "rm -rf lib && NODE_ENV=production babel src/ --out-dir lib --copy-files"
+        "build:babel": "rm -rf lib && NODE_ENV=production babel src/ --out-dir lib --copy-files",
+        "build": "rm -rf lib && rollup -c"
     },
     "peerDependencies": {
         "realm": "^10.10.1"
